changed back from route.mjs to index.mjs
